Refactor image encoding and improve code clarity
- Updated image writing functions to use formatted strings for better readability.
- Added `#[allow(dead_code)]` to the `generate_noisy_image` function to suppress warnings.
- Enhanced various functions across the codebase to utilize more concise syntax and improved error handling.
- Introduced `optimal_compression` flag in the compression run configuration for better control over compression settings.
- Cleaned up code by removing unnecessary conditionals and improving variable handling.

+++ src/compress/lz77.rs
@@ -919,7 +919,7 @@ fn distance_to_symbol(distance: u16) -> (u16, u8) {
     debug_assert!((1..=32768).contains(&distance), "Invalid distance");
 
     if distance < 5 {
-        return (distance as u16 - 1, 0);
+        return (distance - 1, 0);
     }
 
     // For larger distances, use bit manipulation
